izminen update category i style modalka

+++ application/models/admin/RequestCatTwoModels.php
@@ -22,24 +22,30 @@ class RequestCatTwoModels extends CI_Model{
         $config['remove_spaces'] = TRUE;
 
         $this->upload->initialize($config);
-        $this->upload->do_upload('img');
-        $image_data = $this->upload->data();
+        
+        
+        if ( !$this->upload->do_upload('img_cattwo'))
+        {
+            $file_name = "";
+        }
+        else
+        {
+            $image_data = $this->upload->data();
+            $config['image_library'] = 'gd2';
+            $config['source_image'] = $image_data['full_path'];
+            $config['new_image'] = APPPATH . '../assets/upload/thumb/cat_two';
+            $config['maintain_ratio'] = TRUE;
+            $config['width'] = 50;
+            $config['height'] = 50;
 
-        //thumb
-        $config['image_library'] = 'gd2';
-        $config['source_image'] = $image_data['full_path'];
-        $config['new_image'] = APPPATH.'../assets/upload/thumb/cat_two';
-        $config['maintain_ratio'] = TRUE;
-        $config['width'] = 50;
-        $config['height'] = 50;
+            $this->image_lib->initialize($config);
+            $this->image_lib->resize();
+            $name = $this->upload->data();
+            $file_name = $name['file_name'];
+            
 
-        $this->image_lib->initialize($config);
-        $this->image_lib->resize();
-
-        //in DB
-        $name = $this->upload->data();
-        $img = $name['file_name'];
-        return $img;
+        }
+        return $file_name;
     }
 
 
@@ -64,8 +70,9 @@ class RequestCatTwoModels extends CI_Model{
         $title = $this->input->post('title', true);
         $content = $this->input->post('content', true);
         $id_cat = $this->input->post('id', true);
-//        $img = $this->image();
-        if (!$_FILES){
+        $img_name = $this->input->post('img_name');
+        $img = $this->image();
+        if ($img == ""){
             $data = array(
                 'title'=>$title,
                 'description'=>$content,
@@ -73,16 +80,17 @@ class RequestCatTwoModels extends CI_Model{
             );
         }
         else{
-            $img = $this->image();
             $data = array(
                 'title'=>$title,
                 'description'=>$content,
                 'image_name'=>$img,
                 'id'=>$id_cat
             );
-            $img_name = $this->input->post('img_name');
-            unlink('assets/upload/images/cat_two/'. $img_name);
-            unlink('assets/upload/thumb/cat_two/'. $img_name);
+            
+            if($img_name){
+                unlink('assets/upload/images/cat_two/'. $img_name);
+                unlink('assets/upload/thumb/cat_two/'. $img_name);
+            }
         }
         $this->db->where('id', $id_cat);
         $this->db->update('category_two', $data);
